Refactor endpoint mapping in new-services.store.ts
