+++ GUI/src/services/service-builder.ts
@@ -1,5 +1,8 @@
 import axios from "axios";
-import { jsonToYml } from "resources/api-constants";
+import { Edge, Node } from "reactflow";
+import { createNewService, jsonToYml } from "resources/api-constants";
+import useServiceStore from "store/new-services.store";
+import useToastStore from "store/toasts.store";
 import { RawData, Step, StepType } from "types";
 import { EndpointData, EndpointEnv, EndpointType, EndpointVariableData } from "types/endpoint";
 
@@ -113,71 +116,72 @@ const saveEndpointInfo = async (
       }
     )
     .then(console.log)
-    .catch(console.log);
+    .catch(console.log)
 };
 
-const saveEndpointConfig = async (
-  endpoint: EndpointType,
-  env: EndpointEnv,
-  endpointName: string,
-  data: EndpointData,
-) => {
-  const headers = rawDataIfVariablesMissing(
-    endpoint,
-    "headers",
-    env,
-    assignEndpointVariables(env, "headers", endpoint.headers)
-  );
-  const body = rawDataIfVariablesMissing(endpoint, "body", env, assignEndpointVariables(env, "body", endpoint.body));
-  const params = rawDataIfVariablesMissing(
-    endpoint,
-    "params",
-    env,
-    assignEndpointVariables(env, "params", endpoint.params)
-  );
-  const steps = new Map();
-  const variables: { [key: string]: string } = {};
-  assignValues(headers, "headers", variables);
-  assignValues(body, "body", variables);
-  assignValues(params, "params", variables);
-  steps.set("prepare_step", {
-    assign: variables,
-  });
-  steps.set("combine_step", {
-    assign: {
-      sensitive: `\${new Map([${typeof headers === "string"
-          ? `["headers", headers]`
-          : `["headers", new Map([${Object.keys(headers ?? {}).map(
-            (h) => `["${h.replaceAll("__", ".")}", headers_${h}]`
-          )}])]`
-        }, ${typeof body === "string"
-          ? `["body", body]`
-          : `["body", new Map([${Object.keys(body ?? {}).map((b) => `["${b.replaceAll("__", ".")}", body_${b}]`)}])]`
-        }, ${typeof params === "string"
-          ? `["params", params]`
-          : `["params", new Map([${Object.keys(params ?? {}).map(
-            (p) => `["${p.replaceAll("__", ".")}", params_${p}]`
-          )}])]`
-        }])}`,
-    },
-  });
-  steps.set("return_value", { wrapper: false, return: "${sensitive}" });
-  const result = Object.fromEntries(steps.entries());
-  await axios
-    .post(
-      jsonToYml(),
-      { result },
-      {
-        params: {
-          location: `/Ruuter/POST/services/endpoints/configs/${data.isCommon ? "common/" : ""}${endpointName}-${
-            env === EndpointEnv.Live ? "prod" : "test"
-          }-configs.yml`,
-        },
-      }
-    )
-    .then(console.log)
-    .catch(console.log);
-};
+  const saveEndpointConfig = async (
+    endpoint: EndpointType,
+    env: EndpointEnv,
+    endpointName: string,
+    data: EndpointData,
+  ) => {
+    const headers = rawDataIfVariablesMissing(
+      endpoint,
+      "headers",
+      env,
+      assignEndpointVariables(env, "headers", endpoint.headers)
+    );
+    const body = rawDataIfVariablesMissing(endpoint, "body", env, assignEndpointVariables(env, "body", endpoint.body));
+    const params = rawDataIfVariablesMissing(
+      endpoint,
+      "params",
+      env,
+      assignEndpointVariables(env, "params", endpoint.params)
+    );
+    const steps = new Map();
+    const variables: { [key: string]: string } = {};
+    assignValues(headers, "headers", variables);
+    assignValues(body, "body", variables);
+    assignValues(params, "params", variables);
+    steps.set("prepare_step", {
+      assign: variables,
+    });
+    steps.set("combine_step", {
+      assign: {
+        sensitive: `\${new Map([${typeof headers === "string"
+            ? `["headers", headers]`
+            : `["headers", new Map([${Object.keys(headers ?? {}).map(
+                (h) => `["${h.replaceAll("__", ".")}", headers_${h}]`
+              )}])]`
+          }, ${typeof body === "string"
+            ? `["body", body]`
+            : `["body", new Map([${Object.keys(body ?? {}).map((b) => `["${b.replaceAll("__", ".")}", body_${b}]`)}])]`
+          }, ${typeof params === "string"
+            ? `["params", params]`
+            : `["params", new Map([${Object.keys(params ?? {}).map(
+              (p) => `["${p.replaceAll("__", ".")}", params_${p}]`
+            )}])]`
+          }])}`,
+      },
+    });
+    steps.set("return_value", { wrapper: false, return: "${sensitive}" });
+    const result = Object.fromEntries(steps.entries());
+    
+    await axios
+      .post(
+        jsonToYml(),
+        { result },
+        {
+          params: {
+            location: `/Ruuter/POST/services/endpoints/configs/${data.isCommon ? "common/" : ""}${endpointName}-${
+              env === EndpointEnv.Live ? "prod" : "test"
+            }-configs.yml`,
+          },
+        }
+      )
+      .then(console.log)
+      .catch(console.log)
+  };
 
 
 const rawDataIfVariablesMissing = (
@@ -265,30 +269,15 @@ const assignEndpointVariables = (
   return result;
 };
 
-export async function saveEndpoints(endpoints: EndpointData[], name: string, onSuccess: (e: any) => void, onError: (e: any) => void) {
-  if (endpoints.length === 0) return;
-  const setupEndpoints: EndpointData[] | undefined = endpoints;
-  const elements: Step[] = [];
-  setupEndpoints?.forEach((endpoint) => {
-    const selectedEndpoint = endpoint.definedEndpoints.find((e) => e.isSelected);
-    if (!selectedEndpoint) return;
-    elements.push({
-      id: elements.length,
-      label:
-        endpoint.name.trim().length > 0
-          ? endpoint.name
-          : `${selectedEndpoint.methodType.toUpperCase()} ${selectedEndpoint.url}`,
-      type: StepType.UserDefined,
-      data: endpoint,
-    });
-  });
-  for (const endpoint of elements) {
+export async function saveEndpoints(steps: Step[], name: string, onSuccess: (e: any) => void, onError: (e: any) => void) {
+  for (const endpoint of steps) {
     if (!endpoint?.data) continue;
     const selectedEndpointType = endpoint.data.definedEndpoints.find((e) => e.isSelected);
     if (!selectedEndpointType) continue;
 
-    const endpointName = `${name.replaceAll(" ", "_")}-${(endpoint.data.name.trim().length ?? 0) > 0 ? endpoint.data?.name.replaceAll(" ", "_") : endpoint.data?.id
-      }`;
+    const endpointName = `${name.replaceAll(" ", "_")}-${
+      (endpoint.data.name.trim().length ?? 0) > 0 ? endpoint.data?.name.replaceAll(" ", "_") : endpoint.data?.id
+    }`;
     for (const env of [EndpointEnv.Live, EndpointEnv.Test]) {
       await saveEndpointInfo(selectedEndpointType, env, endpointName, endpoint.data);
     }
@@ -375,6 +364,7 @@ export async function saveEndpoints(endpoints: EndpointData[], name: string, onS
       next: "end",
     });
     const result = Object.fromEntries(steps.entries());
+
     await axios
       .post(
         jsonToYml(),
@@ -391,3 +381,326 @@ export async function saveEndpoints(endpoints: EndpointData[], name: string, onS
       .catch(onError);
   }
 }
+
+
+
+
+export const saveFlow = async (steps:Step[], name: string, edges: Edge<any>[], nodes: Node<any, string | undefined>[], onSuccess: (e: any) => void, onError: (e: any) => void, t: any, serviceId: string, description: string, isCommon: boolean) => {
+  try {
+    await saveEndpoints(steps, name, onSuccess, onError);
+    const allRelations: any[] = [];
+    // find regular edges 1 -> 1
+    edges.forEach((edge) => {
+      const node = nodes.find((node) => node.id === edge.source);
+      const followingNode = nodes.find((node) => node.id === edge.target);
+      if (!node) return;
+      if (node.data.stepType === StepType.Textfield && node.data.message === undefined) {
+        throw new Error(t("toast.missing-textfield-message") ?? "Error");
+      }
+      if (
+        node.data.stepType === StepType.OpenWebpage &&
+        (node.data.link === undefined || node.data.linkText === undefined)
+      ) {
+        throw new Error(t("toast.missing-website") ?? "Error");
+      }
+
+      if (
+        node.data.stepType === StepType.FileGenerate &&
+        (node.data.fileName === undefined || node.data.fileContent === undefined)
+      ) {
+        throw new Error(t("toast.missing-file-generation") ?? "Error");
+      }
+
+      if (node.data.stepType === StepType.Input || followingNode?.type === "placeholder") {
+        if (!allRelations.includes(node.id)) allRelations.push(node.id);
+        return;
+      }
+      allRelations.push(`${edge.source}-${edge.target}`);
+    });
+    // find finishing nodes
+    edges.forEach((edge) => {
+      const current = edges.find((lastEdge) => lastEdge.source === edge.source);
+      const nextStep = edges.find((lastEdge) => lastEdge.source === edge.target);
+      if (!nextStep && current?.type !== "placeholder") allRelations.push(edge.target);
+    });
+
+    const finishedFlow = new Map();
+    finishedFlow.set("get_secrets", {
+      call: "http.get",
+      args: {
+        url: `${process.env.REACT_APP_API_URL}/secrets-with-priority`,
+      },
+      result: "secrets",
+    });
+    allRelations.forEach((r) => {
+      const [parentNodeId, childNodeId] = r.split("-");
+      const parentNode = nodes.find((node) => node.id === parentNodeId);
+      if (
+        !parentNode ||
+        parentNode.type !== "customNode" ||
+        [StepType.Rule, StepType.RuleDefinition].includes(parentNode.data.stepType)
+      )
+        return;
+
+      const childNode = nodes.find((node) => node.id === childNodeId);
+      const parentStepName = `${parentNode.data.stepType}-${parentNodeId}`;
+      if (parentNode.data.stepType === StepType.Input) {
+        if (parentNode.data.rules === undefined) {
+          throw new Error(t("toast.missing-client_input-rules") ?? "Error");
+        }
+
+        const clientInput = `ClientInput_${parentNode.data.clientInputId}`;
+        const clientInputName = `${clientInput}-step`;
+        finishedFlow.set(parentStepName, getTemplate(steps, parentNode, clientInputName, `${clientInput}-assign`));
+        finishedFlow.set(`${clientInput}-assign`, {
+          assign: {
+            [clientInput]: `\${${clientInput}_result.input}`,
+          },
+          next: `${clientInput}-switch`,
+        });
+
+        finishedFlow.set(
+          `${clientInput}-switch`,
+          getSwitchCase(
+            edges
+              .filter((e) => e.source === parentNodeId)
+              .map((e) => {
+                const node = nodes.find((node) => node.id === e.target);
+                if (!node) return e.target;
+                const matchingRule = parentNode.data?.rules?.find(
+                  (_: never, i: number) => `rule ${i + 1}` === node.data.label
+                );
+                const followingNode = nodes.find(
+                  (n) => n.id === edges.find((edge) => edge.source === node.id)?.target
+                );
+                return {
+                  case:
+                    matchingRule && !["Yes", "No"].includes(matchingRule?.condition)
+                      ? `\${${matchingRule.name.replace("{{", "").replace("}}", "")} ${matchingRule.condition} ${
+                          matchingRule.value
+                        }}`
+                      : `\${${clientInput} == ${node.data.label === "rule 1" ? '"Yes"' : '"No"'}}`,
+                  nextStep:
+                    followingNode?.type === "customNode"
+                      ? `${followingNode.data.stepType}-${followingNode.id}`
+                      : "service-end",
+                };
+              })
+          )
+        );
+        return;
+      }
+      return finishedFlow.set(
+        parentStepName,
+        getTemplate(steps, parentNode, parentStepName, childNode ? `${childNode.data.stepType}-${childNodeId}` : childNodeId)
+      );
+    });
+    finishedFlow.set("service-end", {
+      wrapper: false,
+      return: "",
+    });
+    const result = Object.fromEntries(finishedFlow.entries());
+    await axios
+      .post(
+        createNewService(),
+        {
+          name,
+          serviceId,
+          description,
+          type: "POST",
+          content: result,
+          isCommon,
+          structure: /* location?.state ?? */ {},
+        },
+        {
+          params: {
+            location: "/Ruuter/POST/services/tests.yml",
+          },
+        }
+      )
+      .then(onSuccess)
+      .catch(onError);
+  } catch (e: any) {
+    onError(e);
+    useToastStore.getState().error({
+      title: t("toast.cannot-save-flow"),
+      message: e?.message,
+    });
+  }
+};
+
+const getMapEntry = (value: string) => {
+  const secrets = useServiceStore.getState().secrets;
+
+  const parts = value.replace("{{", "").replace("}}", "").split(".");
+  const key = value.replace("{{", '"').replace("}}", '"');
+  if ([...(secrets?.prod ?? []), ...(secrets?.test ?? [])].includes(value)) {
+    return `[${key}, secrets.response.body.${parts.join(".")}]`;
+  }
+  if (!value.includes("ClientInput")) parts.splice(1, 0, "response", "body");
+  return `[${key}, ${parts.join(".")}]`;
+};
+
+const getNestedPreDefinedRawVariables = (data: { [key: string]: any }, result: string[]) => {
+  Object.keys(data).forEach((k) => {
+    if (typeof data[k] === "object") {
+      return getNestedPreDefinedRawVariables(data[k], result);
+    }
+    if (typeof data[k] === "string" && data[k].startsWith("{{")) {
+      result.push(getMapEntry(data[k]));
+    }
+  });
+};
+
+const getNestedPreDefinedEndpointVariables = (variable: EndpointVariableData, result: string[]) => {
+  const variableData = variable.type === "schema" ? variable.schemaData : variable.arrayData;
+  if (variableData instanceof Array) {
+    (variableData as EndpointVariableData[]).forEach((v) => {
+      if (["schema", "array"].includes(v.type)) getNestedPreDefinedEndpointVariables(v, result);
+
+      if (v.value && v.value.startsWith("{{")) result.push(getMapEntry(v.value));
+      if (v.testValue && v.testValue.startsWith("{{")) result.push(getMapEntry(v.testValue));
+    });
+  }
+};
+
+const getPreDefinedEndpointVariables = (data?: { variables: EndpointVariableData[]; rawData: RawData }): string[] => {
+  if (!data) return [];
+  const result: string[] = [];
+  data.variables.forEach((v) => {
+    if (!v.value) getNestedPreDefinedEndpointVariables(v, result);
+
+    if (v.value && v.value.startsWith("{{")) result.push(getMapEntry(v.value));
+    if (v.testValue && v.testValue.startsWith("{{")) result.push(getMapEntry(v.testValue));
+  });
+  try {
+    getNestedPreDefinedRawVariables(JSON.parse(data.rawData?.value ?? "{}"), result);
+    getNestedPreDefinedRawVariables(JSON.parse(data.rawData?.testValue ?? "{}"), result);
+  } catch (_) {}
+
+  return result;
+};
+
+const getSwitchCase = (conditions: any[]) => {
+  return {
+    switch: conditions.map((c) => {
+      return {
+        condition: c.case,
+        next: c.nextStep,
+      };
+    }),
+  };
+};
+
+const getTemplate = (steps: Step[], node: Node, stepName: string, nextStep?: string) => {
+  const data = getTemplateDataFromNode(node);
+  if (node.data.stepType === StepType.UserDefined) {
+    return {
+      ...getDefinedEndpointStep(steps, node),
+      next: nextStep ?? "service-end",
+    };
+  }
+  return {
+    template: `templates/${data?.templateName}`,
+    requestType: "post",
+    body: data?.body,
+    result: data?.resultName ?? `${stepName}_result`,
+    next: nextStep ?? "service-end",
+  };
+};
+
+const getTemplateDataFromNode = (node: Node): { templateName: string; body?: any; resultName?: string } | undefined => {
+  if (node.data.stepType === StepType.Auth) {
+    return {
+      templateName: "tara",
+      resultName: "TARA",
+    };
+  }
+  if (node.data.stepType === StepType.Textfield) {
+    return {
+      templateName: "send-message-to-client",
+      body: {
+        message: `${node.data.message?.replace("{{", "${").replace("}}", "}")}`,
+      },
+    };
+  }
+  if (node.data.stepType === StepType.Input) {
+    return {
+      templateName: "client-input",
+      resultName: `ClientInput_${node.data.clientInputId}_result`,
+    };
+  }
+  if (node.data.stepType === StepType.FileGenerate) {
+    return {
+      templateName: "file-generate",
+      body: {
+        fileName: node.data.fileName ?? "",
+        fileContent: node.data.fileContent ?? "",
+      },
+    };
+  }
+  if (node.data.stepType === StepType.FileSign) {
+    return {
+      templateName: "siga",
+      body: {
+        type: "smart_id",
+        country: "EE",
+      },
+      resultName: "SiGa",
+    };
+  }
+  if (node.data.stepType === StepType.FinishingStepRedirect) {
+    return {
+      templateName: "direct-to-cs",
+      body: {
+        message: node.data.message ?? "",
+      },
+    };
+  }
+  if (node.data.stepType === StepType.FinishingStepEnd) {
+    return {
+      templateName: "end-conversation",
+      body: {
+        message: node.data.message ?? "",
+      },
+    };
+  }
+  if (node.data.stepType === StepType.OpenWebpage) {
+    return {
+      templateName: "open-webpage",
+      body: {
+        link: node.data.link ?? "",
+        linkText: node.data.linkText ?? "",
+      },
+    };
+  }
+};
+
+const getDefinedEndpointStep = (steps: Step[], node: Node) => {
+  const endpoint = steps.find((e) => e.label === node.data.label)?.data;
+  const selectedEndpoint = endpoint?.definedEndpoints.find((e) => e.isSelected);
+  if (!selectedEndpoint || !endpoint) {
+    return {
+      return: "",
+    };
+  }
+  return {
+    call: `http.post`,
+    args: {
+      url: `${
+        process.env.REACT_APP_API_URL
+      }/services/endpoints/${selectedEndpoint.methodType.toLowerCase()}-${name}-${
+        (endpoint.name.trim().length ?? 0) > 0 ? endpoint.name : endpoint.id
+      }?type=prod`,
+      body: {
+        headers: `\${new Map([${getPreDefinedEndpointVariables(selectedEndpoint.headers)}])}`,
+        body: `\${new Map([${getPreDefinedEndpointVariables(selectedEndpoint.body)}])}`,
+        params: `\${new Map([${getPreDefinedEndpointVariables(selectedEndpoint.params)}])}`,
+      },
+      params: {
+        type: "prod",
+      },
+    },
+    result: (endpoint.name.trim().length ?? 0) > 0 ? endpoint.name : endpoint.id,
+  };
+};
